+++ MoreHead/Morehead.cs
@@ -16,7 +16,7 @@ public class Morehead : BaseUnityPlugin
 {
     private const string PluginGuid = "Mhz.REPOMoreHead";
     private const string PluginName = "MoreHead";
-    private const string PluginVersion = "1.3.5";
+    private const string PluginVersion = "1.3.6";
     // 单例实例
     public static Morehead? Instance { get; private set; }